Add Alltrucksresult , companiesin contracts, and service

+++ TruckManagementWeb.Core/Models/Reports/TruckMonthReportViewModel.cs
@@ -7,7 +7,7 @@ namespace TruckManagementWeb.Core.Models.Reports
     {
         public int Id {  get; set; }
         public string PlateNumber { get; set; } = string.Empty;
-        public decimal TotalKilometers { get; set; }
+        public int TotalKilometers { get; set; }
         public decimal TotalEuros { get; set; }
         public decimal TotalExpenses { get; set; }
         public IEnumerable<TripViewModel> Trips { get; set; } = new List<TripViewModel>();
